tour: insert intermediate points
to allow for 5m-resolution of gps-position while following the tour.
when leaving the tour, the user is lead back to the most recently visited point,
which might be far away (for long segments with low point-density).
by inserting intermediate points with a distance of 5m, the last point is 5 meters away at max.